+++ R/solution.R
@@ -1,4 +1,0 @@
-solution <- function(x) {
-  if(!inherits(x, "POMDP")) stop("x needs to be a POMDP object!")
-  x$solution
-}