+++ R/solver_output.R
@@ -1,4 +1,0 @@
-solver_output <- function(x) {
-  cat(x$solver_output, sep = "\n")
-  invisible(x$solver_output)
-}